Fix backend and cert origins on dev & workers.dev.
Update the logic for computing fallback_url and cert_url so that it works on
all of `wrangler dev`, `wrangler publish` with workers_dev = true, and
`wrangler publish` with workers_dev = false. Updated the config-generator per
these changes.

Also made private_key_jwk optional in the SXG_CONFIG, and fixed the error
message.

HTML_HOST is not optional because `wrangler publish` doesn't remove vars. If
the user first publishes with workers_dev = true, and then deletes
wrangler.toml and republishes with workers_dev = false, the HTML_HOST should be
'' so it can override the old value.