Make platform-specific landing pages. (#74)
Add the necessary info to the READMEs for cloudflare_worker and fastly_compute
so that they can be effective landing pages. This introduces some doc
duplication, but it seems preferable over making the visitor jump
back-and-forth between docs.

Also add a stub README for the library. Before making this a crate, we should
expand this README and add doc comments to the code.