Upgrade nom to 7.0.0 and pem to 1.0.0.
For nom, the biggest change is the removal of the macros and migration to
function combinators. The second biggest change is the way that alt handles
Incomplete; luckily this only affects internal APIs, and hence required changes
to unit tests only.

For pem, the return type changed from Vec<u8> to Result<Vec<u8>>.