Promote <link rel=preload> tags to Link headers.
Use HTMLRewriter to find tags of the form `<link rel=preload href=... as=...>`
and copy them into `Link` HTTP header directives. This, in turn, enables the
code in header_integrity.rs to add the corresponding `rel=allowed-alt-sxg`
directives so that these subresources can prefetched from SXG referrers.

The HTML must be explicitly UTF-8, either via `Content-Type` header or `<meta
charset>`. Otherwise, we're not sure that HTMLRewriter will parse it correctly.
This could probably be expanded to other ASCII-compatible encodings if we did
some testing. It could probably not include UTF-16, so if we wanted to remove
the requirement of an explicit character encoding, we'd need to add a
content-based sniffer like compact_det_enc or chardetng.

It runs on a copy of the ReadableStream (via body.tee()) to minimize risk of
breakage, since we're only using HTMLRewriter to parse the HTML, not rewrite
it.

Because it is implemented in JS, it is limited to cloudflare_worker. A future
commit will implement it for fastly_compute via the lol_html Rust library.