Extract an HTMLProcessor interface.
Extracts the <link> tag scraping from promoteLinkTagsToHeaders into an
HTMLProcessor class, renaming the function to processHTML and leaving it to do
the UTF-8 verification and orchestration of given processors.

This improves readability by separating concerns, and will help reduce
duplication in a future commit that processes data-sxg-only attributes.

There is a possible performance regression, though I don't know if it's
significant. In order to accommodate the possibility that an HTMLProcessor
intentionally makes a modification (as will be the case in the future commit),
the processed HTML is buffered rather than discarded as it streams in. But for
the majority of cases where there is no modification, this buffer is a wasted
8MB allocation.