Fix streams test flake.
`writer.close()` was consistently throwing an exception `TypeError: Cannot
close a ERRORED writable stream`. The exception was only sometimes being
noticed by Jasmine because it was not being `await`ed, and thus was falling
back to Jasmine's global detector for unhandled promise rejections. This only
succeeded if the test harness wasn't destructed before the `close()` call
finished.

Changed it not to call `close()` in this case, and to always call `await` to
make future such test issues non-flaky.