Reduce clones of response body. (#348)
Change process_html to take and return an Arc<HttpResponse> instead of an
HttpResponse. This allows it to either return the original or the transformed
payload, while the caller keeps a reference to the original. Now, the response
body should only be copied if transformation occurs (the document has a
`<meta charset=utf-8>`).

This copy is necessary in the event that create_signed_exchange errors. In that
case, http_server will process_html(is_sxg=false) on the original.

Alternatively, we could change process_html to output information to allow
process_html(is_sxg=false) to be applied on top of process_html(is_sxg=true).
For instance, when handling template[data-sxg-only] if is_sxg=true, it could
insert comments before and after the content. If is_sxg=false, it could look
for those comments and reinsert the template tag. Left a TODO for that.

Another alternative could be to predict whether signing would fail before
transforming. This would be imperfect -- e.g. transformation may increase or
decrease the size of the payload across the 8MB boundary -- but maybe a close
enough approximation. Whether this is an optimization or a pessimization
(effectively signing twice) depends on how often signing will fail (e.g. due to
cache-control headers).